fix(web): route deploy-skew + RC-exhaustion errors through Sentry beforeSend
Two families of client errors spiked after a deploy. Neither is a logic
bug, but both escaped the existing safeguards.

Webpack deploy-skew: a stale tab requests a code-split composer chunk
(gif picker, video upload) whose module id changed in the new build, so
the webpack runtime is handed an undefined module factory. isDeploySkewError
already matches the Chrome/Firefox/Safari phrasings and the render
boundaries reclassify these to a single low-severity group, but when the
failing chunk is an async dynamic import the rejection surfaces as
onunhandledrejection and is captured directly by Sentry's global handler,
bypassing the boundaries and spawning a fresh error-level issue per deploy.

RC exhaustion: when a user's own account is out of Resource Credits the
chain rejects the broadcast. The vote slider awaited the broadcast with no
catch, so the rejection escaped as unhandled; and the existing filter keyed
on "Insufficient Resource Credits", which the raw RPCError never contains.

- Extract the skew matchers into a React-free utils/deploy-skew, re-exported
  from service-worker-recovery so existing imports are unchanged, so the
  client Sentry config can share them.
- Extract beforeSend into utils/sentry-before-send and reclassify skew
  rejections into the same warning/deploy-skew-auto-recovered group the
  render boundaries use, instead of letting them spawn new issues.
- Broaden the RC filter to all phrasings (has_mana / not enough RC mana /
  Please wait to transact / needs N RC), independent of the handled flag.
- Catch the rejected vote broadcast at the call site and surface the
  friendly insufficient-RC toast via formatError instead of an unhandled
  rejection.
- Add beforeSend regression specs for the skew reclassification and RC drop.
