+++ apps/web/sentry.client.config.ts
@@ -4,7 +4,7 @@
 
 import * as Sentry from "@sentry/nextjs";
 import appPackage from "./package.json";
-import { isEmptyCaptureEvent } from "./src/utils/sentry-empty-capture";
+import { beforeSend } from "./src/utils/sentry-before-send";
 
 // Defer Sentry initialization until after first user interaction or 5s idle.
 // This moves ~1s of JS evaluation off the critical rendering path.
@@ -44,142 +44,10 @@ const SENTRY_CONFIG: Sentry.BrowserOptions = {
     /chrome-extension:\/\//
   ],
 
-  beforeSend(event) {
-    // Drop value-less captures — captureException(null/undefined/"") produces a
-    // synthetic exception with no message and no stack frames (zero actionable
-    // info, all grouped as the "<unknown>" issue). Only empty + frame-less
-    // events are dropped, so real errors (which always carry a stack or a
-    // message) are never lost. See sentry-empty-capture for the exact rule.
-    if (isEmptyCaptureEvent(event)) {
-      return null;
-    }
-
-    const exceptionType = event.exception?.values?.[0]?.type ?? "";
-    const message = event.exception?.values?.[0]?.value ?? "";
-    const stackStr = JSON.stringify(
-      event.exception?.values?.[0]?.stacktrace?.frames ?? []
-    );
-
-    // ECENCY-NEXT-1AA7: an injected page script (pageHook.js / __mm__updateUrl)
-    // JSON.stringify'ing a DOM node. Gate on the injected-script frame so a real
-    // app-side circular-stringify bug (different frames) is still reported.
-    if (
-      /Converting circular structure to JSON/i.test(message) &&
-      /pageHook|__mm__/i.test(stackStr)
-    ) {
-      return null;
-    }
-
-    // ECENCY-NEXT-XA0: "Insufficient Resource Credits" is an expected Hive
-    // condition surfaced as a HANDLED toast. Drop only the handled capture — a
-    // future unhandled path that surfaces the same text is still reported.
-    if (
-      /Insufficient Resource Credits/i.test(message) &&
-      event.exception?.values?.[0]?.mechanism?.handled !== false
-    ) {
-      return null;
-    }
-
-    // AbortController-induced timeouts (TimeoutError / AbortError) ship
-    // with no stack frames, so we can't tell which fetch is at fault.
-    // Walk recent breadcrumbs for the last in-flight fetch URL and tag
-    // the event so we can correlate timeouts to specific endpoints.
-    // Trigger only on the canonical AbortController error names plus the
-    // standard "signal timed out" phrase — a broader /aborted/i match
-    // would also tag unrelated paths (transaction aborts, stream aborts).
-    if (
-      (exceptionType === "TimeoutError" ||
-        exceptionType === "AbortError" ||
-        /signal timed out/i.test(message)) &&
-      !event.tags?.timeoutUrl
-    ) {
-      const crumbs = event.breadcrumbs ?? [];
-      for (let i = crumbs.length - 1; i >= 0; i--) {
-        const c = crumbs[i];
-        const rawUrl = (c.data as { url?: string } | undefined)?.url;
-        if (
-          (c.category === "fetch" || c.category === "xhr") &&
-          typeof rawUrl === "string"
-        ) {
-          // Strip query/hash before tagging — request URLs can carry
-          // tokens or identifiers we don't want as searchable telemetry.
-          let safeUrl: string;
-          try {
-            const parsed = new URL(rawUrl, window.location.origin);
-            safeUrl = `${parsed.origin}${parsed.pathname}`;
-          } catch {
-            safeUrl = rawUrl.split(/[?#]/)[0] ?? "";
-          }
-          if (safeUrl) {
-            event.tags = { ...event.tags, timeoutUrl: safeUrl.slice(0, 200) };
-          }
-          break;
-        }
-      }
-    }
-
-    // React SSR streaming hydration issue triggered by browser extensions
-    // ($RS is React's internal resumable script marker)
-    // Covers old Chrome ≤116 ("Cannot read property 'parentNode' of null"),
-    // new Chrome 117+ ("Cannot read properties of null (reading 'parentNode')"),
-    // Firefox ("can't access property \"parentNode\", a is null"), and
-    // Safari ("null is not an object (evaluating 'a.parentNode')") phrasings.
-    const isParentNodeNull =
-      message.includes("reading 'parentNode'") ||
-      message.includes("property 'parentNode'") ||
-      message.includes('"parentNode"') ||
-      /null is not an object \(evaluating '[a-z]\.parentNode'\)/.test(message);
-    if (isParentNodeNull && stackStr.includes("$RS")) {
-      return null;
-    }
-
-    // Firefox-specific iframe teardown error following a React hydration
-    // mismatch (#418) on profile pages. Symptom — when hydration fails,
-    // React tears down the tree and an iframe's contentWindow becomes null
-    // while some downstream code still tries to access .document on it.
-    // V8-based engines produce a different message for the same access
-    // (`Cannot read properties of null (reading 'document')`) so matching
-    // on the Firefox phrasing alone is engine-specific, but we *also*
-    // require browser=Firefox and a profile-page URL so unrelated iframe
-    // bugs in other surfaces aren't silently dropped. Sample 1% to keep
-    // trend visibility.
-    // TODO(hydration): investigate Firefox-only hydration drift on profile
-    // pages (`/@user/...`) — likely a locale/Date/Intl mismatch or a value
-    // that differs between SSR and the first client render.
-    // event.contexts is loosely typed by @sentry/types; coerce to string.
-    const browserName = String(event.contexts?.browser?.name ?? "");
-    const url = String(event.request?.url ?? "");
-    if (
-      message.includes("can't access property \"document\"") &&
-      message.includes("contentWindow is null") &&
-      /Firefox/i.test(browserName) &&
-      /\/@/.test(url) &&
-      Math.random() > 0.01
-    ) {
-      return null;
-    }
-
-    // Chrome botnet traffic - null/undefined document access from synthetic handler
-    if (
-      /Cannot read properties of (null|undefined) \(reading 'document'\)/.test(
-        message
-      ) &&
-      stackStr.includes("HTMLDocument.c")
-    ) {
-      return null;
-    }
-
-    // Network issues corrupting JS chunk downloads
-    if (
-      /^SyntaxError/.test(event.exception?.values?.[0]?.type ?? "") &&
-      /Unexpected (end of input|token)/.test(message) &&
-      /chunks?[/\\-]/.test(stackStr)
-    ) {
-      return null;
-    }
-
-    return event;
-  }
+  // The single source of truth for client-side event filtering and
+  // reclassification (deploy-skew, RC exhaustion, extension noise, timeouts).
+  // Extracted to src/utils/sentry-before-send so it stays unit-testable.
+  beforeSend
 };
 
 if (typeof window !== "undefined" && process.env.NODE_ENV === "production") {
